collapsed unnecessary component hierarchy
